Added support for necromancer while polling stats from endpoint Added possibility to simulate until X basestats Added possibility to simulate arena gold gains Added companions to experimental dungeon simulator If there is only one configuration saved it is now loaded automatically on page enter Fixed an error if character polled from SFTools had no guild

+++ SFSimulator.Core/Fights/Configurations/ClassConfiguration.cs
@@ -3,6 +3,6 @@
     public sealed class ClassConfiguration
     {
         public double WeaponMultiplier { get; init; }
-        public int HealthMultiplier { get; init; }
+        public double HealthMultiplier { get; init; }
     }
 }
